Add a simpler array based instantiation interface
No need to create the different decay vertices in the right order any
more.  Instead, pass an array with all the particle masses and the
number of particles (length of the array).

NB: Only one level of decay vertices are supported at the moment.

The array should look something like these:

1. Bs → Ds*(Dsγ)ρ(ππ)
   double masses[7] = {Bs, Ds*, ρ, Ds, γ, π, π};
2. Bs → Ds*(Dsγ)π
   double masses[5] = {Bs, Ds*, π, Ds, γ};
3. Bs → Ds*(Dsπ)π
   double masses[5] = {Bs, Ds*, π, Ds, π};