Bug fix "smart" interface to TwoBodyDecayGen
The constructor was not aware of the number of decay vertices in the
decay tree.  When only one of the 2 daughters is unstable and decays
subsequently, it was adding superfluous daughter particles for the
stable daughter.

NB: the higher level decays need not be sorted, but the last decay is