Fix decay tree traversal
Logic as outlined in 9b0313ef was flawed.  There is only one kind of
node as the class TwoBodyDecayGen represents the mother and two
daughters (the doubles, not the pointers to the subsequent decay
vertices).  So only the concept of a leaf node makes sense.

         O
        / \
       /   \    (no subsequent daughters)
      d1   d2

Current implementation of decay tree traversal: Follow each daughter
node, until a leaf node is found.  Return the depth of the recursive
call, so in the current function scope 0 increment by 1 as we go up
the tree.  Once a leaf node is found, branch off the search to the
other daughters one level above (recursion level = 1).